Close sifr workspace phase
