+++ crates/sifr/tests/verification/diagnostics/decimal_invalid_literal/baselines/check-compact.stderr.txt
@@ -1,3 +1,3 @@
 summary: 1 error(s), 0 warning(s), 0 note(s), 0 help item(s)
-error [SIFR-TYPE-0001] [E2501] Decimal() received invalid exact literal '12.34.56' (x1)
+error [SIFR-TYPE-0001] [main] [E2501] Decimal() received invalid exact literal '12.34.56' (x1)
   url: https://sifr.dev/docs/errors/SIFR-TYPE-0001
