+++ crates/sifr/tests/verification/diagnostics/decimal_invalid_literal/baselines/check-human.stderr.txt
@@ -1,1 +1,1 @@
-type error: [E2501] Decimal() received invalid exact literal '12.34.56'
+type error: [main] [E2501] Decimal() received invalid exact literal '12.34.56'
