+++ internal_docs/roadmap.md
@@ -53,7 +53,7 @@ This roadmap is the authoritative execution plan for the current hardening and e
 | 30 | Stdlib Parity (Behavior + Complexity) | completed | [30_reliability_parity_and_performance_budgets.md](./phases/30_reliability_parity_and_performance_budgets.md) | CPython parity governance for stdlib behavior and complexity; milestone_30_4 structural closure completed (2026-03-10) |
 | 31 | Algorithmic Compatibility (LeetCode) | complete | [31_algorithmic_compatibility_and_leetcode_coverage.md](./phases/31_algorithmic_compatibility_and_leetcode_coverage.md) | Corpus-driven compatibility signal, remediation backlog, remediation wave 1, and scorecard/handoff landed on 2026-03-11 with external review sign-off recorded |
 | 31.5 | Ad Hoc Python Source Parity and Builtin Stdlib Surface Closure | in_progress | [issues/ad-hoc-python-source-parity-and-builtin-stdlib-surface.md](../issues/ad-hoc-python-source-parity-and-builtin-stdlib-surface.md), [issues/ad-hoc-python-source-parity-and-builtin-stdlib-surface-execution.md](../issues/ad-hoc-python-source-parity-and-builtin-stdlib-surface-execution.md), [issues/ad-hoc-python-source-parity-extension-waiver-reduction.md](../issues/ad-hoc-python-source-parity-extension-waiver-reduction.md), [issues/ad-hoc-python-source-parity-extension-waiver-reduction-execution.md](../issues/ad-hoc-python-source-parity-extension-waiver-reduction-execution.md), [issues/ad-hoc-structured-data-and-class-surface-parity-expansion.md](../issues/ad-hoc-structured-data-and-class-surface-parity-expansion.md), [issues/ad-hoc-structured-data-and-class-surface-parity-expansion-execution.md](../issues/ad-hoc-structured-data-and-class-surface-parity-expansion-execution.md), [issues/ad-hoc-first-class-bytes-and-binary-surface-foundation.md](../issues/ad-hoc-first-class-bytes-and-binary-surface-foundation.md), [issues/ad-hoc-first-class-bytes-and-binary-surface-foundation-execution.md](../issues/ad-hoc-first-class-bytes-and-binary-surface-foundation-execution.md), [issues/ad-hoc-runtime-and-file-object-parity-expansion.md](../issues/ad-hoc-runtime-and-file-object-parity-expansion.md), [issues/ad-hoc-runtime-and-file-object-parity-expansion-execution.md](../issues/ad-hoc-runtime-and-file-object-parity-expansion-execution.md), [issues/ad-hoc-canonical-iteration-model-and-lazy-parity-closure.md](../issues/ad-hoc-canonical-iteration-model-and-lazy-parity-closure.md), [issues/ad-hoc-canonical-iteration-model-and-lazy-parity-closure-execution.md](../issues/ad-hoc-canonical-iteration-model-and-lazy-parity-closure-execution.md), [issues/ad-hoc-stateful-rng-crypto-and-polish-parity-expansion.md](../issues/ad-hoc-stateful-rng-crypto-and-polish-parity-expansion.md), [issues/ad-hoc-stateful-rng-crypto-and-polish-parity-expansion-execution.md](../issues/ad-hoc-stateful-rng-crypto-and-polish-parity-expansion-execution.md), [issues/ad-hoc-ownership-aware-collection-lowering-and-clone-elision.md](../issues/ad-hoc-ownership-aware-collection-lowering-and-clone-elision.md), [issues/ad-hoc-ownership-aware-collection-lowering-and-clone-elision-execution.md](../issues/ad-hoc-ownership-aware-collection-lowering-and-clone-elision-execution.md), [verification/stdlib/milestone_psp_7_parity_governance_inventory.md](../verification/stdlib/milestone_psp_7_parity_governance_inventory.md) | Root-cause-first closure of builtin/object-model/stdlib parity before Phase 32 continues through sequenced continuation phases. Closed continuations: parity-extension waiver reduction (`wave_psp_ext_1`-`wave_psp_ext_4`, 2026-03-18), structured/class parity (`wave_psp_struct_0`-`wave_psp_struct_4`), bytes foundation (`wave_psp_bytes_0`-`wave_psp_bytes_5`), runtime/file parity (`wave_psp_runtime_0`-`wave_psp_runtime_4`), canonical iteration closure (`wave_psp_iter_fix_0`-`wave_psp_iter_fix_8`), and stateful RNG/crypto closure (`wave_psp_rng_0`-`wave_psp_rng_3`) with phase-level production-grade approval on 2026-03-21. Closed corrective continuation: ownership-aware collection lowering and clone elision (`wave_clone_0` baseline/architecture lock, `wave_clone_1` iterator/comprehension ownership correction, `wave_clone_2` index/slice/star-unpack ownership correction, and `wave_clone_3` generic hardening/regression lock completed on 2026-03-21; wave-closure pass-1/pass-2, milestone-closure pass-1/pass-2, and phase-closure pass-1/pass-2 production-grade reviews approved on 2026-03-21). |
-| 31.6 | Ad Hoc Sifr Workspace Resolution Via `sifr.toml` | ready_to_implement | [issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25.md](../issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25.md), [issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25-execution.md](../issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25-execution.md) | Workspace-rooted user imports for non-`main.sifr` entries, preserving stdlib resolver precedence and enabling shared LeetCode/user helpers without stdlib pollution. |
+| 31.6 | Ad Hoc Sifr Workspace Resolution Via `sifr.toml` | closed | [issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25.md](../issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25.md), [issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25-execution.md](../issues/ad-hoc-sifr-workspace-sifr-toml-import-resolution-2026-04-25-execution.md) | Shipped native `sifr.toml` workspace discovery, workspace-rooted user imports for non-`main.sifr` entries, dotted module materialization, validation fixtures, and the LeetCode helper pilot without stdlib pollution. |
 
 ## Expansion and Distribution Track (Phase 32-41)
